snapshot file data is always written in ordered mode
the next3 mount options "ordered" determines the way that file data pages
are written to disk on user file writes.
it has no effect on the way that snapshot file data pages are written to
disk, which is either in "ordered" mode or sync_dirty_buffer().
in fact, there is no writepage() operation at all in next3_snapfile_aops,
because user write access to snapshots is not permitted.