patch atomic_snapshot_take.patch
seperate last snaphsot on-disk (s_last_snapshot)
from active snaphsot on-disk (s_snapshot_inum).

the on-disk field s_last_snapshot used to be set at the end of snapshot_take(),
so a crash between snapshot_create() and snapshot_take() would leave a floating
snapshot file, which would not be cleaned on next mount and on tune2fs -O ^has_snapshot.

the old field s_last_snapshot was renamed to s_snapshot_inum and a new field named s_last_snapshot
is used to store the head of on-disk snapshots list and is set at the beginning of snapshot_create().
the 'has_snapshot' feature is also set at the beginning of snapshot_create() now.

snapshot_load() starts by loading s_last_snapshot, but if that fails, it continues to load s_snapshot_inum.