check_is_handle_aborted.patch
the practice in ext3 void "delete" functions is to check for is_handle_aborted(handle),
because the only errors expected are journaling errors, which are marked on the transaction.
snapshots don't change that behavior, because snapshot COW errors would also cause the transaction to abort.

changed get_clear_access() to void and documented that it causes transaction to abort on errors.