patch should_cow_data.patch
this patch fixes a critical bug: on partial page write,
the entire block is moved to snaphsot and the new block is initialized to zero
and not to the data of the moved block.

to solve this bug, move-on-write is done only on full page write, when the buffer is marked with buffer_move() state flag.
on partial page write, the data block is copied-on-write, because this is more efficient than the alternative.

as a by product of this change, sqlite database file blocks (in-place partial page writes) will be copied-on-write
and the file will not get fragmented.

move to snapshot still happens on unlink/truncate and full page in-place write (for example in bare metal backup).