patch read_before_move.patch
discard the previous patch should_cow_data, because it had some problems.
instead, read moved block to page cache before moving it to snapshot
on partial page write.

direct_IO is disabled in next3, because it has not been tested properly and may cause bugs.

next3_block_truncate_page() is not called in next3, as a simple solution to protect snapshot
data on last page truncate.