clear_buffer_tracked_read.patch
renamed block device page buffer _bh to bdev_bh.
bdev_bh is attached to a page in the block device address space and is used by the f/s to synchronize tracked readers
of the same block from different snapshot address spaces.
changing the refcount of bdev_bh can be done in any order to clearing bh flags.
the only important thing to maintain is balanced calls to get/put_bh_tracked_reader(bdev_bh).

buffers are attached to pages and have refcount maintained by paging sub-system.
as long as the page is in use, the buffer won't get freed.

I also don't care if bdev_bh gets freed when it's refcount drops to 0, because that would mean
that there is no pending COW to that block, so I have no need for the tracked readers count.

complicated patch this tracked read...