complete_cow_answers.patch
the order of the calls is as they are.I copied it from other places in the code and it is important.
for ordered data:
	unlock_buffer(sbh);
        err = next3_journal_dirty_data(handle, sbh);
        mark_buffer_dirty(sbh);

for writeback data:
	unlock_buffer(sbh);
        mark_buffer_dirty(sbh);

and for synced data write:
	unlock_buffer(sbh);
	mark_buffer_dirty(sbh);
	sync_dirty_buffer(sbh);

complete_cow can do either, depending on @handle and @sync args.

if I could unlock the buffer later in end_pending_cow(), there would be no need for the busy wait on buffer_new()
in test_pending_cow().
the wait_on_buffer() is test_pending_cow() prevent the 2nd busy wait, when there is already a busy wait for tracked readers in
complete_cow() (while buffer is locked).