truncate_page_answers.patch
there are 2 calls to get_blocks_handle() in this fxn.
the 1st one with @create=0 and if the block is not mapped, the code returns from this fxn.
the 2nd call, which I added with @create=1 triggers the move-on-write.
the result of a move-on-write is a 'new' allocated buffer, which has to be treated differently.
after the call to unmap_underlying_metadata() the 'new' flag is no longer needed.