balloc_fixes.patch
the fix_exclude flag signals an exclude bitmap error and prompts to run fsck.
currently, when running fsck, a message is displayed that fixing exclude bitmap
is not supported when snapshots exists and that the only option is to discard 
all snapshots and run fsck again.
in the future, I will implement fsck of exclude bitmap (not very complicated).