fsck_exclude_bitmap.patch
now that fsck of exclude bitmap is implemented, we can also use it to verify there are no
excluded blocks inside the snapshot image.
we fix the exclude bitmap on read through from snapshot to be all zeroes.
we fix copy of snapshot inodes to regular (non-excluded) files on snaphsot take.