avoid unmapping of buffers on journal dirty data list
it is unnecessary and very risky to unmap a buffer on the journal
dirty data list, becasue kjournald may be trying to flush it.
check the buffer_jbd() flag to see if the buffer is journaled.
data buffers can only be journaled as dirty data when working in
data=ordered mode, which is the only mode supported by next3.