added support for extent feature
to enable extent feature on filesystem created with mkfs.nexts do:
> next4 umount
> next4 off
> tune2fs -O extent
> next4 on
> next4 mount

next4 tests script now works well when extent feature is on.

the fixes include:

1] snapshot files are always created as indirect mapped files.

2] update inode set_flags mask.

3] reimplement snapshot_file_huge patch.
extra tripple indirect block are now stored in in-memory i_data[0-2]
instead of in i_data[13-15].
on-disk they are still stored in i_block[0-2].
this is needed so sizeof(i_data[]) is the same as sizeof(i_block[]),
because extents code checks sizeof(i_data[]) in several places.