resolve circular locking dependency
COW of bitmap block during ext4_mb_new_blocks() triggers a lockdep warning:

=======================================================
[ INFO: possible circular locking dependency detected ]
2.6.37-rc2.debug+ #3
-------------------------------------------------------
dd/3506 is trying to acquire lock:
 (&ei->i_data_sem#2/1){++++..},
 	at: [<ffffffffa00bd40b>] next4_map_blocks+0x6b/0x2b0 [next4]

but task is already holding lock:
 (&meta_group_info[i]->alloc_sem#2){++++..},
 	at: [<ffffffffa00e925f>] next4_mb_load_buddy+0xbf/0x3c0 [next4]

which lock already depends on the new lock.
=======================================================

To avoid this warning (and possible deadlock), we avoid taking
down_read(&grp->alloc_sem) on every block group, but the last one
(seperate patch does that).

In this patch, we avoid allocating blocks for COW from last block group,
when there are blocks available in other block groups.

This is a band-aid, which will be later replaced by proper removal of
grp->alloc_sem from the blocks allocation path.