merge from next3: fix hang after large snapshot delete
when large snapshot shrink transaction was restarted, truncate_mutex
was released, although it wasn't taken in snapshot_shrink() and later
it was aquired and never freed.

check if truncate_mutex is locked before releasing and re-aquiring it.

in ext4, replaced the mutex_is_locked() test with explicit
ext4_snapshot_file() test.