jbd2: add the h_cowing flag and debug counters to jbd2_journal_handle struct
The h_cowing flag is used by ext4 snapshots code to mark the task in COWING
state. Other field are used to track COW buffer credits and debug counters.

Merging this patch to kernel 2.6.39, will allow users to test ext4
snapshots as a standalone module (with debug counters enabled), without
the need to patch and install a development kernel.

With CONFIG_JBD2_DEBUG not defined, this patch doesn't change the size of
the struct.

With CONFIG_JBD2_DEBUG defined, the struct grows by 7 int fields, but since
CONFIG_JBD2_DEBUG is mostly defined on development kernels, and since the
size of objects associated with a task is much larger that 7 ints, that
doesn't seem to be a big problem.

Signed-off-by: Amir Goldstein <amir73il@users.sf.net>