Eliminate a useless lookup on MOW.
Lookup on MOW is not needed if EXT4_GET_BLOCK_CREATE is set, because
lookup is finished without EXT4_GET_BLOCK_CREATE and set EXT4_MAP_MOW
if MOW is needed.

EXT4_MAP_MOW indicates that MOW is needed on the requested block.

I modified the patch a bit and added some comments. (Amir)