don't bypass journal if flex_bg feature is set
There are currently problems with the flex_bg feature, because
some places in snapshots code assume that the block bitmap is
in the block group (which is not true for flex_bg).

This fix just avoids potential metadata corruption, but it doesn't
avoid possible OOPS.

There are other problem to solve, because snapshots created with flex_bg
feature do not pass the fsck sanity test.