remove next4 script from ext4-snapshots repo
The script was moved to e2fsprogs-snapshots repo, were
it was renamed to next3 and is installed by install_next4.sh
script to /sbin/next4, along with the .next4 utilities it uses.

Please note that the format of .next4.conf was changed, so you
will have to run 'next4 config' to rewrite the file.