fix blocks count of uninit_bg with exclude_bitmap and flex_bg
ext4_group_used_meta_blocks() counts metadata blocks in the group.
it always counted 1 exclude bitmap, even if exclude bitmap was not
in the group, in the case of flex_bg.

Note that this function doesn't count the number of metadata blocks
in the first flex_bg group correctly, which is why mkfs always initializes
the block bitmap of the first flex_bg group.