limit the time of long snapshot shrink/merge transactions
snapshot shrink/merge transactions are extended/restarted
as many times as needed during the long operation.
in some cases, when the shrink/merge performs very little
metadata updates, it may take a very long time before the
transaction handle is released.
in that situation, other fs writers can block for a very long time.

to fix this in merge, a single transaction is started for every indirect
branch move, so most of the time there is no open handle.
to fix this in shrink, the long transaction is always retsarted when
crossing a block group boundary.