Fix a bug in shrink/merge of snapshots of volume size >= 8TB
next3_snapshot_shrink_blocks(), next3_snapshot_merge_blocks()
and their helper functions were not designed to deal with unsigned
values of blocks count, but they are allowed to
return any number of blocks count > 0 and <= range.
To workaround this problem, never cross the signed/unsigend value
of maxblocks and count on the shrink/merge loop to call again with
(iblock+count, range-count).

Signed-off-by: Amir Goldstein <amir73il@gmail.com>
Signed-off-by: Gilad Shaiak <gilad@ctera.com>