ovl: cow file to snapshot on open for write
With snapshot=/snapshot mount option, ovl_snapshot_d_real() will
recursively call d_real() on the /snapshot dentry, which will
"copy down" the file to the snapshot's upper dir.
Then, ovl_snapshot_d_real() will return its own upper dentry
(not the snapshot's upper dentry), so the stable copy of the file
will remain in /snapshot.

Snapshot dentry may be positive or negative or NULL.
If positive, it may need to be copied down.
If negative, it may need to be explicitly whited out.
Otherwise, the entry is nested inside an already whited out directory,
so need to do nothing about it

Signed-off-by: Amir Goldstein <amir73il@gmail.com>