fsnotify: implement event reporting to super block's root inode
When the flag FS_EVENT_ON_SB is set on a super block's root inode
fsnotify mask, all events on inodes that are on the same super
block are reported to the root inode.

Events reported to root inode carry the flag FS_EVENT_ON_SB,
to distinguish from events that happen to the root inode itself
or to its direct children (FS_EVENT_ON_CHILD).

Hook the root inode reporting in fsnotify() for events of type
FSNOTIFY_EVENT_PATH, because those are the events that are also
supported for mount watches by fanotify.

Signed-off-by: Amir Goldstein <amir73il@gmail.com>