fanotify: support events with data type FSNOTIFY_EVENT_DENTRY
When event data type is FSNOTIFY_EVENT_DENTRY, store the
provided dentry in the event queue along with the mount point
that was used to setup the watch.

This will enable the watcher to read the path of the passed
dentry from /proc/self/fd/X, for example on FAN_ATTRIB events.

Signed-off-by: Amir Goldstein <amir73il@gmail.com>