fanotify: add a super block root watch
Set a watch on the super block's root inode including the
mask bit FAN_EVENT_ON_SB to get notified on the events
of all the inodes on the same super block.

The mark must be set on a root inode and it cannot
be passed along with FAN_MARK_MOUNT, because super block watch
is an inode watch, not a mount watch.

Signed-off-by: Amir Goldstein <amir73il@gmail.com>