fanotify: new init flag FAN_EVENT_INFO_PARENT
The new init flag FAN_EVENT_INFO_PARENT indicates that the
event fd may point to the parent, depending on the event.

FAN_EVENT_INFO_PARENT will also be used to add support for more
events, so it is stored in a new 'flags' field in the group's
private data for future use.

FAN_EVENT_INFO_PARENT can only be set along with the default
notification class.

Signed-off-by: Amir Goldstein <amir73il@gmail.com>