fanotify: report events to sb root with fanotify_file_event_info
Allocate variable length fanotify_file_event_info struct to
report events when FAN_EVENT_ON_SB mask is set on event.

This is needed to provide extra info on the event for the
super block watcher.

An exception is made for permission event which use the
fanotify_perm_event_info struct also when reporting to
super block watcher.

Signed-off-by: Amir Goldstein <amir73il@gmail.com>