ovl: explicit copy up/whiteout requests from snapshot mount
snapshot mount allows changes to overlayfs lower, while maintaining
a stable view in merge of upper+lower.

snapshot mount calls the underlying overlayfs d_real() with
O_RDWR open_flags.
Unlike vfs_open(), snapshot mount can pass non zero open_flags
with non regular dentry as well as with a negative dentry.

d_real() is never called from vfs with non zero open_flags and
non regular/negative dentry, so this change has no effect on existing
overlayfs vfs operations.

The purpose of the call with non regular dentry is a request for copy
up of dir/symlink in preparation of upcoming delete or modify in lower.

The purpose of the call with negative dentry is a request for explicit
whiteout, in preparation for upcoming create in lower.

ovl_do_remove() is altered to only perform explicit whiteout
when called with a negative dentry.

Signed-off-by: Amir Goldstein <amir73il@gmail.com>