ovl: revalidate snapshot mount dentries
Snapshot mount writes rely on testing the value of __snapdentry
to determine if file needs to be copied up by snapshot overlay
that is mounted at __snapmnt.

If __snapmnt is changed while snapshot mount is mounted (i.e.
live snapshot take), the values of __snapdentry in existing
snapshot mount dentries becomes stale and point to dentries
in the old snapshot overlay.

Implement the d_revalidate() operation for snapshot mount to
check the condition of stale __snapdentry value.
When ovl_snapshot_revalidate() detects a stale __snapdentry value,
it will return 0 to indicate that vfs lookup should invalidate the
dentry and call ovl_lookup() again to lookup __snapdentry in the
new __snapmnt.

The expected result is that modifying a file after live snaphot take,
will trigger copy up in the new snapshot overlay.

Signed-off-by: Amir Goldstein <amir73il@gmail.com>