ovl: implement snapshot mount lookup
Snapshot mount lookup is about finding out if a file
has already been copied or whited out in snapshot overlay.

Find a reference to the dentry in the snapshot overlay mount
pointed to by snapshot= mount option.

The snapshot layer lookup in the snapshot overlay mount
falls through to look at the snapshot mount upper layer.

The only 2 reasons for a snapshot dentry to be negative are:
1. entry is not existing in snapshot mount upper layer
2. entry is whiteout in snapshot overlay upper layer

We need to hold a reference to the snapshot overlay negative
dentry, so we can use it to signal the overlay mount to
explicitly whiteout a non existing entry.

Signed-off-by: Amir Goldstein <amir73il@gmail.com>