ovl: implement snapshot mount lookup
Snapshot mount lookup is about finding out if a file has already been
copied up or whited out in the snapshot overlay.

For a given snapshot mount path, find a reference to the dentry with
the same path in the snapshot overlay mount.

When looking up in the overlay snapshot mount, all we need to know
about the snaphot dentry is if it is positive and if it is upper, so
we know if we need to CoW it. If the snapshot dentry is positive, we
don't care if it is dir/non-dir and whether it is opaque/redirect.

The snapshot layer lookup in the snapshot overlay mount falls through
to look at the snapshot mount upper layer, which is the snapshot overlay
lower layer.

The only 2 reasons for a snapshot dentry to be negative are:
1. entry is not existing in neither upper nor lower layers
2. entry is whiteout in snapshot overlay upper layer

We need to hold a reference to the snapshot overlay negative dentry, so
we can use it to signal the overlay mount to explicitly whiteout a non
existing entry.

Signed-off-by: Amir Goldstein <amir73il@gmail.com>