ovl: introduce snapshot mount
Register overalyfs also as file system type 'snapshot'.

A 'snapshot mount' is stacked on top of an upperdir and all the
operations on the snapshot mount pass through to the upperdir.

Later on, the snapshot mount will provide copy on write to
a snapshot storage directory.

A snapshot mount has no lowerdir and there is no copy up,
so there is no need for a workdir.  Fix some places in overlayfs
code that assume the existence of lowerdir and workdir.

The snapshot mount is enabled by the experimental config
option CONFIG_OVERLAY_FS_SNAPSHOT.

Signed-off-by: Amir Goldstein <amir73il@gmail.com>