ovl: reserve stack depth for snapshot overlay
A snapshot mount may be mounted with no snapshot overlay
and remounted later with an underlying snapshot overlay.

We must leave room in stack below the snapshot mount
for that overlay, even if snapshot= mount option is
not provided on the initial mount.

We must also enforce the reserved stack depth when
replacing the snapshot overlay on remount.

Signed-off-by: Amir Goldstein <amir73il@gmail.com>