ovl: invalidate overlay dentries after delete in lowerdir
The snapshot overlay is a special use case for an overlayfs mount.
Usually, making modification to the lowerdir of an overlayfs mount
while the overlay is mounted may result in undefined behavior.

In the special case of a snapshot overlay mount that is protected
by a snapshot mount, we know when modifications are made to lowerdir.
When entries are deleted in the snapshot overlay lowerdir, we need to
invalidate the snapshot overlay dentries that reference these stale
lower entries.

We identify a lowerdir delete operation by the fact that the snapshot
mount dentry is unhashed after the operation and then we also make sure
to unhash the respective snapshot overlay dentry.

Signed-off-by: Amir Goldstein <amir73il@gmail.com>