ovl: cow file to snapshot on open for write
On open for write of file in a snapshot mount, ovl_snapshot_d_real() will
recursively call d_real() on the snapshot overlay dentry, which will
"copy down" the file to the snapshot overlay's upper dir.
Then, ovl_snapshot_d_real() will return its own upper dentry
(not the snapshot overlay's upper dentry), so the stable copy of the file
will remain in the snapshot overlay upper dir.

A snapshot overlay dentry reference in a snapshot mount dentry may be
positive or negative or NULL.
If positive, it may need to be copied up ini the snapshot overlay.
If negative, it may need to be whited out in the snapshot overlay.
Otherwise, the entry is nested inside an already whited out directory,
so need to do nothing about it

Signed-off-by: Amir Goldstein <amir73il@gmail.com>