ovl: handle nesting under new snapshot mount directory
When snapshot mount entries are nested under a new directory,
that ancestor snapshot dentry may be non-dir or negative and
the nested entries used to have NULL snapshot dentires.

Instead of storing a NULL snapshot dentry, point the snapshot
dentry to the snapshot overlay root to indicate this special case.
This is needed for access to snapshot overlay sb, which will be needed
for ovl_snapshot_revalidate().

Signed-off-by: Amir Goldstein <amir73il@gmail.com>