ovl: add support for verify_lower mount option
When overlayfs is mounted with option 'verify_lower', a directory inode
found in lower layer by name or by redirect_dir is verified against the
file handle of the copy up origin that is stored in the upper layer.

The 'verify_lower' option should not be used after copying layers,
because the new lower directory inodes would fail verification.

Internally, 'verify_lower' is implemented as an alias to mount option
'verify_dir=<OVL_VERIFY_LOWER>', where OVL_VERIFY_LOWER is a bitmask
with a single defined flag - __OVL_VERIFY_MERGE. That bitmask is going
to be extended with more optional flags in the future.

Signed-off-by: Amir Goldstein <amir73il@gmail.com>