ovl: relax same fs constrain for constant st_ino
For the case of all layers not on the same fs, use the copy up
origin st_ino for non-dir, if we know it.

This guaranties constant and system-wide unique st_ino/st_dev
for non-dir across copy up. Like the same fs case, st_ino of
non-dir is also persistent. Unlink the same fs case, st_dev is
not persistent.

Signed-off-by: Amir Goldstein <amir73il@gmail.com>