ovl: decode lower dir on failure to verify origin fh
With mount option 'verify_lower', on lookup of a merge dir dentry, if
lower dir does not match the origin fh stored in upper dir, or if no
lower dir was found by name, try to decode the stored origin fh and
use the result as the merge dir lower dentry.

This allows an overlay mount to cope better with changes to lower layer,
like rename and remove of lower merge dir.

A lower merge directory followed by origin fh cannot be followed down
again by name, so practically, 'verify_lower' is not able to cope well
with lower layer changes when there is more than a single lower layer.

Signed-off-by: Amir Goldstein <amir73il@gmail.com>